+++ src/table/cells/withSelections.jsx
@@ -1,15 +1,19 @@
 import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
+import handleCellFocus from './handle-cell-focus';
 import { selectCell } from '../selections-utils';
 import { getSelectionStyle } from '../styling-utils';
 
 export default function withSelections(CellComponent) {
   const HOC = (props) => {
     const { cell, selState, selDispatch, styling, focusedCell, setFocusedCell, rootElement, pageInfo, page } = props;
-    const handleMouseUp = (evt) =>
+    const handleMouseUp = (evt) => {
       cell.isDim &&
-      evt.button === 0 &&
-      selectCell(cell, selState, selDispatch, focusedCell, setFocusedCell, rootElement, pageInfo, page, evt);
+        evt.button === 0 &&
+        handleCellFocus(cell, focusedCell, setFocusedCell, rootElement, pageInfo, page) &&
+        selectCell(cell, selState, selDispatch, focusedCell, setFocusedCell, rootElement, pageInfo, page, evt);
+    };
+
     const selectionStyling = useMemo(() => getSelectionStyle(styling, cell, selState), [styling, cell, selState]);
 
     return <CellComponent {...props} styling={selectionStyling} onMouseUp={handleMouseUp} />;
@@ -23,7 +27,7 @@ export default function withSelections(CellComponent) {
     focusedCell: PropTypes.array.isRequired,
     setFocusedCell: PropTypes.func.isRequired,
     rootElement: PropTypes.object.isRequired,
-    pageInfo: PropTypes.array.isRequired,
+    pageInfo: PropTypes.object.isRequired,
     page: PropTypes.number.isRequired,
   };
